Added Left
FM - Added the getLeft that will take an index position and location the left child for that position. Still need to hanle the case when not found. Also, hidding the base array that is generated when Heap in created. The Heap should be an interface. The users should not be able to access the array.

Signed-off-by: Frank McAuley <frank.mcauley@gmail.com>